+++ src/index.ts
@@ -2,4 +2,7 @@ export { CsvReporter } from './reporters/csv';
 export { JsonReporter } from './reporters/json';
 export { PrettyReporter } from './reporters/pretty';
 export { GatherReporter } from './reporters/gather';
+export * from './middleware/grep';
 export * from './runner';
+export * from './options';
+export { RunFunction, IBenchmarkCase } from './suite';