+++ src/suite.ts
@@ -1,83 +1,66 @@
-import { Suite } from 'benchmark';
-import { Benchmark, IReporter } from './reporters';
 import { MaybeAsync, IDeferred, returnsPromiseLike, runMaybeAsync } from './async';
+import { Options } from './options';
+import { Middleware } from './runner';
+import Benchmark from 'benchmark';
 
 /**
  * Method used for running benchmarks. Can be stubbed in tests.
  */
-export type RunFunction = (name: string, options: IOptions) => Promise<void>;
+export type RunFunction = (name: string, options: Options) => Promise<void>;
 
 const defaultRunFn: RunFunction = (name, options) => {
-  const suite = new Suite();
-  suite.add(name, options);
-  const prom = new Promise<void>(resolve => suite.on('complete', resolve));
-  suite.run();
+  const bench = new Benchmark(name, options.fn!, options);
+  const prom = new Promise<void>(resolve => bench.on('complete', resolve));
+  bench.run();
   return prom;
 };
-
-/**
- * Benchmark options. A subset of those that Benchmark.js has, and better typed.
- */
-export interface IOptions {
-  defer?: boolean;
-  delay?: number;
-  initCount?: number;
-  maxTime?: number;
-  minSamples?: number;
-  minTime?: number;
-  name?: string;
-  onComplete?: (event: { target: Benchmark }) => void;
-  onCycle?: (event: { target: Benchmark }) => void;
-  onStart?: (event: { target: Benchmark }) => void;
-  onError?: (event: { target: Benchmark }) => void;
-  onReset?: (event: { target: Benchmark }) => void;
-  setup?: MaybeAsync;
-  teardown?: MaybeAsync;
-  fn?: (deferred: IDeferred) => void;
-}
 
 export interface IBenchmarkCase {
   name: string;
   fn: MaybeAsync;
-  options?: IOptions;
+  middleware: ReadonlyArray<Middleware>;
+  options?: Options;
 }
 
 /**
  * Installs a callback or promise-returning benchmark in the bench
  * options. This is awkward with their API, but it is what it is.
  */
-const installBenchFn = (fn: MaybeAsync, options: IOptions) => {
+const installBenchFn = (fn: MaybeAsync, options: Options) => {
   if (fn.length > 0) {
-    options.defer = true;
-    options.fn = (deferred: IDeferred) => {
-      fn(err => (err ? deferred.reject(err) : deferred.resolve(undefined)));
-    };
-  } else if (returnsPromiseLike(fn)) {
-    options.defer = true;
-    options.fn = (deferred: IDeferred) => {
-      fn()!
-        .then(() => deferred.resolve(undefined))
-        .catch(err => deferred.reject(err));
-    };
-  } else {
-    options.fn = fn as () => void;
+    return options.merge({
+      defer: true,
+      fn: (deferred: IDeferred) => {
+        fn(err => (err ? deferred.reject(err) : deferred.resolve(undefined)));
+      },
+    });
   }
+
+  if (returnsPromiseLike(fn)) {
+    return options.merge({
+      defer: true,
+      fn: (deferred: IDeferred) => {
+        fn()!
+          .then(() => deferred.resolve(undefined))
+          .catch(err => deferred.reject(err));
+      },
+    });
+  }
+
+  return options.merge({ fn: fn as () => void });
 };
 
-/**
- * Composes a function that adds a call to the `additional` function after
- * the original one, if it exists.
- */
-export const compose = <Args extends unknown[]>(
-  original: ((...fn: Args) => void) | undefined,
-  additional: (...fn: Args) => void,
-) =>
-  original
-    ? (...args: Args) => {
-        original(...args);
-        additional(...args);
-      }
-    : additional;
+const runMiddleware = (
+  bench: Readonly<IBenchmarkCase>,
+  stack: ReadonlyArray<Middleware>,
+  offset = 0,
+): Promise<void> => {
+  if (stack.length === offset) {
+    return Promise.resolve();
+  }
+
+  return stack[offset](bench, b => runMiddleware(b, stack, offset + 1));
+};
 
 /**
  * A benchmark suite. Wraps around the basic Benchmark.js suites, since they
@@ -99,26 +82,28 @@ export class MatchaSuite {
   /**
    * Runs all benchmarks.
    */
-  public async run(reporter: IReporter): Promise<void> {
+  public async run(): Promise<void> {
     for (const bench of this.benches) {
-      const { setup, teardown, ...options } = bench.options ?? {};
-      if (setup) {
-        await runMaybeAsync(setup as MaybeAsync);
-      }
+      await runMiddleware(bench, [
+        ...bench.middleware,
+        async bench => {
+          const options = bench.options ?? Options.empty;
+          if (options.setup) {
+            await runMaybeAsync(options.setup);
+          }
 
-      installBenchFn(bench.fn, options);
+          const benchOptions = options.assign({
+            setup: undefined,
+            teardown: undefined,
+          });
 
-      options.onStart = compose(options.onStart, event => reporter.onStartCycle(event.target));
-      options.onComplete = compose(options.onComplete, event =>
-        reporter.onFinishCycle(event.target),
-      );
-      await this.runFn(bench.name, options);
+          await this.runFn(bench.name, installBenchFn(bench.fn, benchOptions));
 
-      if (teardown) {
-        await runMaybeAsync(teardown as MaybeAsync);
-      }
+          if (options.teardown) {
+            await runMaybeAsync(options.teardown);
+          }
+        },
+      ]);
     }
-
-    reporter.onComplete();
   }
 }